Add sending an email via kafka to API (#140)
* Add functionality to send and email via kafka

Add EmailService to create the data for an email and send it to kafka. This is part of migrating of the use of kafka in lfp-appeals-frontend to lfp-appeals-api

BI-8916

* Updates to fix Spring dependency issues

BI-8916

* Updated email content to match what is sent from the front-end

BI-8916

* Add external email and update content

Added the environment-reader-library to pom.xml. This is used when determining the appropriate region email to send the internal email to in the EmailService.
Created enum Region. Used in above
Update email content to allow the template to display the content correctly

BI-8916

* Add functionality to get the company name

Add api-sdk-manager-java-library and private-api-sdk-java to the pom.
Add CompanyProfileSevice to retrieve a companies profile.
Add ApiClientService to get the CH private SDK used by the CompanyProfileService
Update EmailService to use the CompanyProfileService
Add ServiceException to handle exceptions when sending a message to kafka

BI-8916

* Add unit tests and additional logging

BI-8916

* Update and fix broken tests

Renamed AppealControllerTest_POST to AppealControllerPostTest to follow naming conventions.
Update bootstrap-servers value in the test/resources/application.yml.
Add bean EnvironmentReader to ApplicationConfig. This now means it's managed by Spring so therefore removed the @Autowired annotation from EmailService.
Removed EmailSendMessageProducerIntegrationTest as shouldn't have been committed as it was not needed
Add some extra structured logging and fix some spelling mistakes

BI-8916

* Update code for sonar fixes

Add CompanyProfileServiceTest to cover tests of CompanyProfileService
Update CompanyProfileService by removing the throw of ServiceException and updating declaration of uri to use var as Java 10 local variable type inference

BI-8916

* Add and update tests for code coverage

Add LoggingUtilsTest and EmailSendKafkaProducerTest
Add additional tests to EmailSendMessageProducerTest
Remove commented out code from TestUtils

BI-8916

* Update pom.xml

Add mockito-inline to test static methods in LoggingUtilsTest

Bi-8916

* Update exception handling

Split out the exception handling in EmailSendMessageProducer. This is because the InterruptedException needs to either be rethrown or re-interrupted
Removed unused imports in EmailSendMessageProducerTest

BI-8916

* Fix some code smells

Update code smells. In Java 10/11 there is local type inference so update instances where sonar suggested changing to var

BI-8916

* Remove unused imports