Prevent duplicate appeal records
Before creating an appeal, added a check to see if an appeal for the company number/reference exists. If there are duplicates the first record is updated with the current date/time and the reason passed in updated.
Add new test for updating an appeal
Updated some of the logging calls to use the penalty reference as the context.

BI-11981