Bump spring boot version to 2.7.16 (#220)
There was an original dependabot PR, https://github.com/companieshouse/lfp-appeals-api/pull/168, would have bump spring-boot-starter-web from 2.4.3 to 2.5.12 for a critical vulnerability. This was superseded by a couple more dependabot PRs to go to version 3..0.0. Bumping to 3.0.0 caused build failures so for now I have bumped up to the last version 2 of spring boot, 2.7.16.
Bumped version of api-sdk-java as per Martin Westacott message.
Updated expected message for one of the tests in AppealControllerPostTest as it was failing.

DEPT-1979